pyproject.toml
[tool.poetry]
name = "repl-nix-ip-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_ip_project"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
ipapi = "^1.0.4"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
